Style the floating action button like twitter also making it functional.
