+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
     sdk: flutter
   http: ^1.3.0
   flutter_dotenv: ^5.2.1
+  flutter_speed_dial: ^7.0.0
 
 dev_dependencies:
   flutter_test:
